+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 We have a new DataBag (based on the old DataSift/Stone BaseObject), which comes with full dot.notation.support :)
 
+* Containers\DataBag added
 * Exceptions\E4xx_NoSuchIndex added
 * Exceptions\E4xx_NoSuchProperty added
 * Exceptions\E4xx_UnsupportedType added